pyproject.toml
[project]
name = "jira-refinement-agent"
version = "0.1.0"
description = "Jira-native AI refinement agent — turns rough tickets into sprint-ready specs"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.126.0",
    "uvicorn[standard]>=0.32.0",

    # Config & validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # HTTP client
    "httpx>=0.28.0",

    # MCP client SDK (connects to mcp-atlassian server)
    "mcp>=1.0.0",

    # LLM
    "openai>=1.50.0",

    # YAML config
    "pyyaml>=6.0.2",

    # Logging
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
